updated dev.rake to look like doc

+++ lib/tasks/dev.rake
@@ -3,17 +3,16 @@ desc "Fill the database tables with some sample data"
 # end
 
 task sample_data: :environment do
-  # starting = Time.now
-
   p "Creating sample data"
+  starting = Time.now
 
-  # if Rails.env.development?
+  if Rails.env.development?
     FollowRequest.delete_all
     Comment.delete_all
     Like.delete_all
     Photo.delete_all
     User.delete_all
-  # end
+  end
 
   12.times do
     name = Faker::Name.first_name
